Refactor: Move logstash conf updates into new facade file
