+++ facade/logstash.go
@@ -1,0 +1,61 @@
+// Copyright 2016 The Serviced Authors.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package facade
+
+import (
+	"github.com/zenoss/glog"
+
+	"github.com/control-center/serviced/dao"
+	"github.com/control-center/serviced/datastore"
+	"github.com/control-center/serviced/domain/servicetemplate"
+	"github.com/control-center/serviced/isvcs"
+)
+
+type reloadLogstashContainer func(ctx datastore.Context, f FacadeInterface) error
+
+var LogstashContainerReloader reloadLogstashContainer = reloadLogstashContainerImpl
+
+// writeLogstashConfiguration takes all the available
+// services and writes out the filters section for logstash.
+// This is required before logstash startsup
+func writeLogstashConfiguration(templates map[string]servicetemplate.ServiceTemplate) error {
+	// FIXME: eventually this file should live in the DFS or the config should
+	// live in zookeeper to allow the agents to get to this
+	if err := dao.WriteConfigurationFile(templates); err != nil {
+		return err
+	}
+	return nil
+}
+
+// Anytime the available service definitions are modified
+// we need to restart the logstash container so it can write out
+// its new filter set.
+// This method depends on the elasticsearch container being up and running.
+func reloadLogstashContainerImpl(ctx datastore.Context, f FacadeInterface) error {
+	templates, err := f.GetServiceTemplates(ctx)
+	if err != nil {
+		glog.Errorf("Could not write logstash configuration: %s", err)
+		return err
+	}
+	if err := writeLogstashConfiguration(templates); err != nil {
+		glog.Errorf("Could not write logstash configuration: %s", err)
+		return err
+	}
+	glog.V(2).Infof("Starting logstash container")
+	if err := isvcs.Mgr.Notify("restart logstash"); err != nil {
+		glog.Errorf("Could not start logstash container: %s", err)
+		return err
+	}
+	return nil
+}
